+++ src/components/tab/indicator.module.css.d.ts
@@ -8,10 +8,14 @@
 
 export const root: string;
 
-export const tabList: string;
+export const fullWidth: string;
 
-export const vertical: string;
+export const active: string;
 
-export const tab: string;
+export const bottom: string;
 
-export const indicator: string;
+export const top: string;
+
+export const left: string;
+
+export const right: string;