+++ src/components/tabs/tabs.tsx
@@ -10,8 +10,11 @@ import classNames from 'classnames';
 import React from 'react';
 
 import {OverridableComponent, OverridableProps, mergeClassesProps, chain} from '../../utils';
+import {IndicatorProps, IndicatorPlacement, Indicator} from '../tab';
 
+import {Orientation, Variant, Size} from './common-tabs-props';
 import * as styles from './tabs.module.css';
+import {PositionStyles, getPositionStyles} from './utils';
 
 /** Tabs props */
 export interface Props {
@@ -21,11 +24,24 @@ export interface Props {
     onValueChange?: (value: any) => void;
 
     /** Tabs orientation (default: "horizontal") */
-    orientation?: 'horizontal' | 'vertical';
+    orientation?: Orientation;
     /** Style of tabs (default: "ghost") */
-    variant?: 'ghost' | 'solid';
+    variant?: Variant;
     /** Size of tabs (default: "normal") */
-    size?: 'normal' | 'small' | 'tiny';
+    size?: Size;
+    /** Animation of the tab indicator(default: true) */
+    animated?: boolean;
+
+    /** Props of tabs wrapper */
+    tabListProps?: React.HTMLAttributes<HTMLDivElement>;
+    /**
+     * Indicator placement
+     * Default value for horizontal tabs: "bottom"
+     * Default value for vertical tabs: "right"
+     */
+    indicatorPlacement?: IndicatorPlacement;
+    /** Props of tab indicator */
+    indicatorProps?: IndicatorProps;
 }
 
 interface TypeMap {
@@ -44,49 +60,128 @@ export const Tabs: OverridableComponent<TypeMap> = React.forwardRef<
     OverridableProps<TypeMap>
 >(function TabsComponent(props, ref) {
     const {
-        value: tabsValue,
+        value: selectedValue,
         onValueChange,
 
         component: Component = 'div',
         orientation = 'horizontal',
         variant = 'ghost',
         size = 'normal',
+        animated = true,
+
+        tabListProps = {},
+        indicatorPlacement: outlinePosition,
+        indicatorProps = {},
+
         classes,
-        children,
+        children: childrenList,
         ...restProps
     } = mergeClassesProps(props, styles);
 
+    const valueToIndex = new Map();
+    const [mounted, setMounted] = React.useState(false);
+    const tabListRef = React.useRef<HTMLDivElement>(null);
+    const [indicatorStyles, setIndicatorStyles] = React.useState<PositionStyles | null>(null);
+
+    const indicatorPlacement = React.useMemo(() => {
+        if (outlinePosition) {
+            return outlinePosition;
+        }
+
+        if (orientation === 'vertical') {
+            return 'right';
+        }
+
+        return 'bottom';
+    }, [outlinePosition, orientation]);
+
+    const children: React.ReactElement[] = React.Children.toArray(childrenList).filter(
+        React.isValidElement
+    );
+
+    React.useEffect(() => {
+        setMounted(true);
+    }, []);
+
+    React.useEffect(() => {
+        if (!tabListRef.current) {
+            return undefined;
+        }
+
+        const selectedTabNode = tabListRef.current.children[valueToIndex.get(selectedValue)];
+        if (!selectedTabNode) {
+            setIndicatorStyles(null);
+
+            return undefined;
+        }
+
+        const tabIndicatorStyles = getPositionStyles(
+            tabListRef.current,
+            selectedTabNode,
+            indicatorPlacement,
+            variant
+        );
+
+        setIndicatorStyles(tabIndicatorStyles);
+    }, [indicatorPlacement, variant, size, orientation, selectedValue, childrenList]);
+
     return (
-        <Component
-            {...restProps}
-            ref={ref}
-            className={classNames(classes.root, {
-                [classes.vertical]: orientation === 'vertical'
-            })}
-        >
-            {React.Children.map(children, (child, childIndex) => {
-                if (!React.isValidElement(child)) {
-                    return null;
-                }
+        <Component {...restProps} ref={ref} className={classes.root}>
+            <div
+                role="tablist"
+                {...tabListProps}
+                ref={tabListRef}
+                className={classNames(classes.tabList, {
+                    [classes.vertical]: orientation === 'vertical'
+                })}
+            >
+                {children.map((child, childIndex) => {
+                    const {
+                        onClick,
+                        value,
+                        indicatorProps: tabIndicatorProps = {},
+                        ...restChildProps
+                    } = child.props;
 
-                const {onClick, value, ...restChildProps} = child.props;
-                const childValue = value ?? childIndex;
+                    const childValue = value ?? childIndex;
+                    const selected = childValue === selectedValue;
 
-                const handleClick = React.useCallback(() => {
-                    if (onValueChange) {
-                        onValueChange(childValue);
-                    }
-                }, [onValueChange, childValue]);
+                    valueToIndex.set(childValue, childIndex);
 
-                return React.cloneElement(child, {
-                    value: childValue,
-                    onClick: chain(handleClick, onClick),
-                    selected: childValue === tabsValue,
-                    variant,
-                    size,
-                    ...restChildProps
-                });
-            })}
+                    const handleClick = React.useCallback(() => {
+                        if (onValueChange) {
+                            onValueChange(childValue);
+                        }
+                    }, [onValueChange, childValue]);
+
+                    return React.cloneElement(child, {
+                        className: animated && classes.tab,
+                        value: childValue,
+                        onClick: chain(handleClick, onClick),
+                        selected,
+                        variant,
+                        size,
+                        classes: {
+                            indicator: animated && classes.indicator
+                        },
+                        indicatorPlacement,
+                        indicatorProps: {
+                            ...tabIndicatorProps,
+                            selected: !animated && selected
+                        },
+                        ...restChildProps
+                    });
+                })}
+            </div>
+
+            {animated && mounted && indicatorStyles && (
+                <Indicator
+                    {...indicatorProps}
+                    active
+                    className={classes.indicator}
+                    style={indicatorStyles}
+                />
+            )}
         </Component>
     );
 });